--- pages/api/get_initial_seed.js
@@ -1,2 +1,0 @@
-// lat long
-// return string 'summary'
